Fix overload positional args and infer from defaults
