+++ mypy/typeops.py
@@ -8,8 +8,7 @@ NOTE: These must not be accessed from mypy.nodes or mypy.types to avoid import
 from __future__ import annotations
 
 import itertools
-from collections import defaultdict
-from typing import Any, Callable, Iterable, List, Sequence, TypeVar, cast
+from typing import Any, Iterable, List, Sequence, TypeVar, cast
 
 from mypy.copytype import copy_type
 from mypy.expandtype import expand_type, expand_type_by_instance
@@ -25,7 +24,6 @@ from mypy.nodes import (
     FuncDef,
     FuncItem,
     OverloadedFuncDef,
-    OverloadPart,
     StrExpr,
     TypeInfo,
     Var,
@@ -65,7 +63,6 @@ from mypy.types import (
     flatten_nested_unions,
     get_proper_type,
     get_proper_types,
-    is_unannotated_any,
 )
 from mypy.typevars import fill_typevars
 
@@ -1124,82 +1121,6 @@ def separate_union_literals(t: UnionType) -> tuple[Sequence[LiteralType], Sequen
             union_items.append(item)
 
     return literal_items, union_items
-
-
-def infer_impl_from_parts(
-    impl: OverloadPart,
-    types: list[CallableType],
-    fallback: Instance,
-    named_type: Callable[[str, list[Type]], Type],
-):
-    impl_func = impl if isinstance(impl, FuncDef) else impl.func
-    # infer the types of the impl from the overload types
-    arg_types: dict[str | int, list[Type]] = defaultdict(list)
-    ret_types = []
-    for tp in types:
-        for i, arg_type in enumerate(tp.arg_types):
-            arg_name = tp.arg_names[i]
-            if not arg_name:  # if it's positional only
-                if arg_type not in arg_types[i]:
-                    arg_types[i].append(arg_type)
-            else:
-                if arg_name in impl_func.arg_names:
-                    if arg_type not in arg_types[arg_name]:
-                        arg_types[arg_name].append(arg_type)
-                if arg_name and arg_name in impl_func.arg_names:
-                    if arg_type not in arg_types[arg_name]:
-                        arg_types[arg_name].append(arg_type)
-        t = get_proper_type(tp.ret_type)
-        if isinstance(t, Instance) and t.type.fullname == "typing.Coroutine":
-            ret_type = t.args[2]
-        else:
-            ret_type = tp.ret_type
-        if ret_type not in ret_types:
-            ret_types.append(ret_type)
-    res_arg_types = [
-        (
-            UnionType.make_union((arg_types[arg_name_] if arg_name_ else []) + arg_types[i])
-            if arg_kind not in (ARG_STAR, ARG_STAR2)
-            else UntypedType()
-        )
-        for i, (arg_name_, arg_kind) in enumerate(zip(impl_func.arg_names, impl_func.arg_kinds))
-    ]
-
-    ret_type = UnionType.make_union(ret_types)
-
-    if impl_func.is_coroutine:
-        # if the impl is a coroutine, then assume the parts are also, if not need annotation
-        any_type = AnyType(TypeOfAny.special_form)
-        ret_type = named_type("typing.Coroutine", [any_type, any_type, ret_type])
-
-    # use unanalyzed_type because we would have already tried to infer from defaults
-    if impl_func.unanalyzed_type:
-        assert isinstance(impl_func.unanalyzed_type, CallableType)
-        assert isinstance(impl_func.type, CallableType)
-        impl_func.type = impl_func.type.copy_modified(
-            arg_types=[
-                i if not is_unannotated_any(u) else r
-                for i, u, r in zip(
-                    impl_func.type.arg_types, impl_func.unanalyzed_type.arg_types, res_arg_types
-                )
-            ],
-            ret_type=(
-                ret_type
-                if isinstance(
-                    get_proper_type(impl_func.unanalyzed_type.ret_type), (AnyType, NoneType)
-                )
-                else impl_func.type.ret_type
-            ),
-        )
-    else:
-        impl_func.type = CallableType(
-            res_arg_types,
-            impl_func.arg_kinds,
-            impl_func.arg_names,
-            ret_type,
-            fallback,
-            definition=impl_func,
-        )
 
 
 def try_getting_instance_fallback(typ: Type) -> Instance | None:
